+++ module/video.hpp
@@ -3,11 +3,11 @@
 #include "core/constant.hpp" // IWYU pragma: keep
 
 namespace video {
- u8 pixel(s16 x, s16 y);
- void pixel(s16 x, s16 y, u8 color);
- void line(s16 ax, s16 ay, s16 bx, s16 by, u8 color);
- void rect(s16 x, s16 y, s16 width, s16 height, u8 color, bool fill);
- void text(s16 x, s16 y, str text, u8 color, u8 background = 0);
+ u8 pixel(s32 x, s32 y);
+ void pixel(s32 x, s32 y, u8 color);
+ void line(s32 ax, s32 ay, s32 bx, s32 by, u8 color);
+ void rect(s32 x, s32 y, s32 width, s32 height, u8 color, bool fill);
+ void text(s32 x, s32 y, str text, u8 color, u8 background = 0);
 
  void clear(u8 color);
  void flip();
